Rename resource store
