+++ src/app/mod.rs
@@ -2,11 +2,10 @@ pub mod app;
 pub mod window;
 pub mod gui;
 pub mod engine;
-mod resource_store;
+mod image_resource;
 
 pub use self::app::Cen;
 pub use self::window::Window;
 pub use self::gui::TextureKey;
-pub use self::resource_store::ImageFlags;
-pub use self::resource_store::ImageResource;
-pub use self::resource_store::ResourceStore;
+pub use self::image_resource::ImageFlags;
+pub use self::image_resource::ImageResource;
